Move main application code to dream-application crate
Took 20 minutes
