+++ dream-application/src/icons/DirectoryIcon.png
