+++ dream-application/src/icons/ExpandIcon.png
