+++ dream-application/src/icons/PlayIconDark.png
