+++ dream-application/src/icons/SelectIconBlack.png
