+++ dream-application/src/icons/StopIcon.png
