+++ src/lib.rs
@@ -1,6 +1,6 @@
 /**********************************************************************************
  *  Dream is a software for developing real-time 3D experiences.
- *  Copyright (C) 2023 Deepak Ramalignam
+ *  Copyright (C) 2023 Deepak Ramalingam
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -16,825 +16,19 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **********************************************************************************/
 
-pub mod camera;
-pub mod texture;
-
-use dream_editor;
-use std::iter;
-use wgpu::util::DeviceExt;
-
-use winit::{
-    event::*,
-    event_loop::{ControlFlow, EventLoop},
-    window::{Window, WindowBuilder},
-};
-
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
-
-// lib.rs
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-struct Vertex {
-    position: [f32; 3],
-    tex_coords: [f32; 2],
-}
-
-// lib.rs
-impl Vertex {
-    fn desc<'a>() -> wgpu::VertexBufferLayout<'a> {
-        wgpu::VertexBufferLayout {
-            array_stride: std::mem::size_of::<Vertex>() as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &[
-                // position
-                wgpu::VertexAttribute {
-                    offset: 0,
-                    shader_location: 0,
-                    format: wgpu::VertexFormat::Float32x3,
-                },
-                // uv
-                wgpu::VertexAttribute {
-                    offset: std::mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
-                    shader_location: 1,
-                    format: wgpu::VertexFormat::Float32x2,
-                },
-            ],
-        }
-    }
-}
-
-// lib.rs
-const VERTICES: &[Vertex] = &[
-    Vertex {
-        position: [-0.0868241, 0.49240386, 0.0],
-        tex_coords: [0.4131759, 0.00759614],
-    }, // A
-    Vertex {
-        position: [-0.49513406, 0.06958647, 0.0],
-        tex_coords: [0.0048659444, 0.43041354],
-    }, // B
-    Vertex {
-        position: [-0.21918549, -0.44939706, 0.0],
-        tex_coords: [0.28081453, 0.949397],
-    }, // C
-    Vertex {
-        position: [0.35966998, -0.3473291, 0.0],
-        tex_coords: [0.85967, 0.84732914],
-    }, // D
-    Vertex {
-        position: [0.44147372, 0.2347359, 0.0],
-        tex_coords: [0.9414737, 0.2652641],
-    }, // E
-];
-
-const INDICES: &[u16] = &[0, 1, 4, 1, 2, 4, 2, 3, 4];
-
-// We need this for Rust to store our data correctly for the shaders
-#[repr(C)]
-// This is so we can store this in a buffer
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct CameraUniform {
-    // We can't use cgmath with bytemuck directly so we'll have
-    // to convert the Matrix4 into a 4x4 f32 array
-    view_proj: [[f32; 4]; 4],
-}
-
-impl CameraUniform {
-    fn new() -> Self {
-        use cgmath::SquareMatrix;
-        Self {
-            view_proj: cgmath::Matrix4::identity().into(),
-        }
-    }
-
-    fn update_view_proj(&mut self, camera: &camera::Camera) {
-        self.view_proj = camera.build_view_projection_matrix().into();
-    }
-}
-
-struct State {
-    surface: wgpu::Surface,
-    device: wgpu::Device,
-    queue: wgpu::Queue,
-    config: wgpu::SurfaceConfiguration,
-    size: winit::dpi::PhysicalSize<u32>,
-    render_pipeline: wgpu::RenderPipeline,
-    window: Window,
-    egui_wgpu_renderer: egui_wgpu::Renderer,
-    egui_context: egui::Context,
-    egui_winit_state: egui_winit::State,
-    #[allow(dead_code)]
-    demo_app: egui_demo_lib::DemoWindows,
-    #[allow(dead_code)]
-    diffuse_texture: texture::Texture,
-    depth_texture: texture::Texture,
-    depth_texture_egui: texture::Texture,
-    frame_texture: texture::Texture,
-    frame_texture_view: Option<wgpu::TextureView>,
-    play_icon_texture: texture::Texture,
-    file_icon_texture: texture::Texture,
-    directory_icon_texture: texture::Texture,
-    vertex_buffer: wgpu::Buffer,
-    index_buffer: wgpu::Buffer,
-    diffuse_bind_group: wgpu::BindGroup,
-    camera: camera::Camera,
-    camera_uniform: CameraUniform,
-    camera_buffer: wgpu::Buffer,
-    camera_bind_group: wgpu::BindGroup,
-}
-
-impl State {
-    async fn new(window: Window, event_loop: &EventLoop<()>) -> Self {
-        let size = window.inner_size();
-
-        // The instance is a handle to our GPU
-        // BackendBit::PRIMARY => Vulkan + Metal + DX12 + Browser WebGPU
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-            backends: wgpu::Backends::all(),
-            dx12_shader_compiler: Default::default(),
-        });
-
-        // # Safety
-        //
-        // The surface needs to live as long as the window that created it.
-        // State owns the window so this should be safe.
-        let surface = unsafe { instance.create_surface(&window) }.unwrap();
-
-        let adapter = instance
-            .request_adapter(&wgpu::RequestAdapterOptions {
-                power_preference: wgpu::PowerPreference::default(),
-                compatible_surface: Some(&surface),
-                force_fallback_adapter: false,
-            })
-            .await
-            .unwrap();
-
-        let surface_caps = surface.get_capabilities(&adapter);
-        // Shader code in this tutorial assumes an Srgb surface texture. Using a different
-        // one will result all the colors comming out darker. If you want to support non
-        // Srgb surfaces, you'll need to account for that when drawing to the frame.
-        let surface_format = surface_caps
-            .formats
-            .iter()
-            .copied()
-            .filter(|f| f.describe().srgb)
-            .next()
-            .unwrap_or(surface_caps.formats[0]);
-
-        let mut web_gl_limits = wgpu::Limits::downlevel_webgl2_defaults();
-        web_gl_limits.max_texture_dimension_2d = 4096;
-
-        let (device, queue, egui_wgpu_renderer) = adapter
-            .request_device(
-                &wgpu::DeviceDescriptor {
-                    label: None,
-                    features: wgpu::Features::empty(),
-                    // WebGL doesn't support all of wgpu's features, so if
-                    // we're building for the web we'll have to disable some.
-                    limits: if cfg!(target_arch = "wasm32") {
-                        web_gl_limits
-                    } else {
-                        wgpu::Limits::default()
-                    },
-                },
-                None, // Trace path
-            )
-            .await
-            .map(|(device, queue)| {
-                let egui_wgpu_renderer = egui_wgpu::Renderer::new(
-                    &device,
-                    surface_format,
-                    Some(texture::Texture::DEPTH_FORMAT),
-                    1,
-                );
-                return (device, queue, egui_wgpu_renderer);
-            })
-            .unwrap();
-
-        let config = wgpu::SurfaceConfiguration {
-            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-            format: surface_format,
-            width: size.width,
-            height: size.height,
-            present_mode: surface_caps.present_modes[0],
-            alpha_mode: surface_caps.alpha_modes[0],
-            view_formats: vec![],
-        };
-        surface.configure(&device, &config);
-
-        let diffuse_bytes = include_bytes!("happy-tree.png");
-        let diffuse_texture =
-            texture::Texture::from_bytes(&device, &queue, diffuse_bytes, "happy-tree.png").unwrap();
-
-        let texture_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                entries: &[
-                    wgpu::BindGroupLayoutEntry {
-                        binding: 0,
-                        visibility: wgpu::ShaderStages::FRAGMENT,
-                        ty: wgpu::BindingType::Texture {
-                            multisampled: false,
-                            view_dimension: wgpu::TextureViewDimension::D2,
-                            sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                        },
-                        count: None,
-                    },
-                    wgpu::BindGroupLayoutEntry {
-                        binding: 1,
-                        visibility: wgpu::ShaderStages::FRAGMENT,
-                        // This should match the filterable field of the
-                        // corresponding Texture entry above.
-                        ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                        count: None,
-                    },
-                ],
-                label: Some("texture_bind_group_layout"),
-            });
-
-        let diffuse_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &texture_bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::TextureView(&diffuse_texture.view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::Sampler(&diffuse_texture.sampler),
-                },
-            ],
-            label: Some("diffuse_bind_group"),
-        });
-
-        let camera = camera::Camera {
-            // position the camera one unit up and 2 units back
-            // +z is out of the screen
-            eye: (0.0, 1.0, 2.0).into(),
-            // have it look at the origin
-            target: (0.0, 0.0, 0.0).into(),
-            // which way is "up"
-            up: cgmath::Vector3::unit_y(),
-            aspect: config.width as f32 / config.height as f32,
-            fovy: 45.0,
-            znear: 0.1,
-            zfar: 100.0,
-        };
-
-        // in new() after creating `camera`
-
-        let mut camera_uniform = CameraUniform::new();
-        camera_uniform.update_view_proj(&camera);
-
-        let camera_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Camera Buffer"),
-            contents: bytemuck::cast_slice(&[camera_uniform]),
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-        });
-
-        let camera_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                entries: &[wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                }],
-                label: Some("camera_bind_group_layout"),
-            });
-
-        let camera_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &camera_bind_group_layout,
-            entries: &[wgpu::BindGroupEntry {
-                binding: 0,
-                resource: camera_buffer.as_entire_binding(),
-            }],
-            label: Some("camera_bind_group"),
-        });
-
-        let play_icon_texture_bytes = include_bytes!("icons/PlayIcon.png");
-        let play_icon_texture = texture::Texture::from_bytes(
-            &device,
-            &queue,
-            play_icon_texture_bytes,
-            "icons/PlayIcon.png",
-        )
-        .unwrap();
-
-        let file_icon_texture_bytes = include_bytes!("icons/FileIcon.png");
-        let file_icon_texture = texture::Texture::from_bytes(
-            &device,
-            &queue,
-            file_icon_texture_bytes,
-            "icons/FileIcon.png",
-        )
-        .unwrap();
-
-        let directory_icon_texture_bytes = include_bytes!("icons/DirectoryIcon.png");
-        let directory_icon_texture = texture::Texture::from_bytes(
-            &device,
-            &queue,
-            directory_icon_texture_bytes,
-            "icons/FileIcon.png",
-        )
-        .unwrap();
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("Shader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("shader.wgsl").into()),
-        });
-
-        let depth_texture =
-            texture::Texture::create_depth_texture(&device, &config, "depth_texture");
-
-        let depth_texture_egui =
-            texture::Texture::create_depth_texture(&device, &config, "depth_texture_egui");
-
-        let frame_texture =
-            texture::Texture::create_frame_texture(&device, &config, "frame_texture");
-
-        let render_pipeline_layout =
-            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                label: Some("Render Pipeline Layout"),
-                bind_group_layouts: &[&texture_bind_group_layout, &camera_bind_group_layout],
-                push_constant_ranges: &[],
-            });
-
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("Render Pipeline"),
-            layout: Some(&render_pipeline_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[Vertex::desc()],
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: config.format,
-                    blend: Some(wgpu::BlendState {
-                        color: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::SrcAlpha,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                        alpha: wgpu::BlendComponent::OVER,
-                    }),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleList,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: Some(wgpu::Face::Back),
-                // Setting this to anything other than Fill requires Features::POLYGON_MODE_LINE
-                // or Features::POLYGON_MODE_POINT
-                polygon_mode: wgpu::PolygonMode::Fill,
-                // Requires Features::DEPTH_CLIP_CONTROL
-                unclipped_depth: false,
-                // Requires Features::CONSERVATIVE_RASTERIZATION
-                conservative: false,
-            },
-            depth_stencil: Some(wgpu::DepthStencilState {
-                format: texture::Texture::DEPTH_FORMAT,
-                depth_write_enabled: true,
-                depth_compare: wgpu::CompareFunction::Less,
-                stencil: wgpu::StencilState::default(),
-                bias: wgpu::DepthBiasState::default(),
-            }),
-            multisample: wgpu::MultisampleState {
-                count: 1,
-                mask: !0,
-                alpha_to_coverage_enabled: false,
-            },
-            // If the pipeline will be used with a multiview render pass, this
-            // indicates how many array layers the attachments will have.
-            multiview: None,
-        });
-
-        let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Vertex Buffer"),
-            contents: bytemuck::cast_slice(VERTICES),
-            usage: wgpu::BufferUsages::VERTEX,
-        });
-
-        let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Index Buffer"),
-            contents: bytemuck::cast_slice(INDICES),
-            usage: wgpu::BufferUsages::INDEX,
-        });
-
-        let mut egui_winit_state = egui_winit::State::new(&event_loop);
-        egui_winit_state.set_pixels_per_point(window.scale_factor() as f32);
-        let egui_winit_context = egui::Context::default();
-        let demo_app = egui_demo_lib::DemoWindows::default();
-
-        Self {
-            surface,
-            device,
-            queue,
-            size,
-            config,
-            render_pipeline,
-            window,
-            egui_wgpu_renderer,
-            egui_winit_state,
-            egui_context: egui_winit_context,
-            demo_app,
-            diffuse_texture,
-            depth_texture,
-            depth_texture_egui,
-            frame_texture,
-            frame_texture_view: None,
-            play_icon_texture,
-            file_icon_texture,
-            directory_icon_texture,
-            vertex_buffer,
-            index_buffer,
-            diffuse_bind_group,
-            camera,
-            camera_uniform,
-            camera_buffer,
-            camera_bind_group,
-        }
-    }
-
-    pub fn window(&self) -> &Window {
-        &self.window
-    }
-
-    pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
-        if new_size.width > 0 && new_size.height > 0 {
-            self.size = new_size;
-            self.config.width = new_size.width;
-            self.config.height = new_size.height;
-            self.surface.configure(&self.device, &self.config);
-            // resize frame textures
-            self.frame_texture =
-                texture::Texture::create_frame_texture(&self.device, &self.config, "frame_texture");
-            self.depth_texture =
-                texture::Texture::create_depth_texture(&self.device, &self.config, "depth_texture");
-            // resize egui depth texture
-            self.depth_texture_egui = texture::Texture::create_depth_texture(
-                &self.device,
-                &self.config,
-                "depth_texture_egui",
-            );
-        }
-    }
-
-    fn update(&mut self) {}
-
-    fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
-        let texture_view = self
-            .frame_texture
-            .texture
-            .create_view(&wgpu::TextureViewDescriptor::default());
-
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("Render Encoder"),
-            });
-
-        // draw triangle
-        {
-            let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                label: Some("Render Pass"),
-                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                    view: &texture_view,
-                    resolve_target: None,
-                    ops: wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(wgpu::Color {
-                            r: 0.1,
-                            g: 0.2,
-                            b: 0.3,
-                            a: 1.0,
-                        }),
-                        store: true,
-                    },
-                })],
-                depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                    view: &self.depth_texture.view,
-                    depth_ops: Some(wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(1.0),
-                        store: true,
-                    }),
-                    stencil_ops: None,
-                }),
-            });
-
-            // let num_vertices = VERTICES.len() as u32;
-            let num_indices = INDICES.len() as u32;
-
-            render_pass.set_pipeline(&self.render_pipeline);
-            // diffuse texture
-            render_pass.set_bind_group(0, &self.diffuse_bind_group, &[]);
-            // camera
-            render_pass.set_bind_group(1, &self.camera_bind_group, &[]);
-            render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
-            render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
-            // vertex drawing
-            render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
-            render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
-            render_pass.draw_indexed(0..num_indices, 0, 0..1);
-        }
-
-        self.queue.submit(iter::once(encoder.finish()));
-
-        let output_texture_view = self
-            .frame_texture
-            .texture
-            .create_view(&wgpu::TextureViewDescriptor::default());
-
-        self.frame_texture_view = Some(output_texture_view);
-
-        Ok(())
-    }
-
-    fn render_egui(&mut self) -> Result<(), wgpu::SurfaceError> {
-        let output = self.surface.get_current_texture()?;
-        let view = output
-            .texture
-            .create_view(&wgpu::TextureViewDescriptor::default());
-
-        let input = self.egui_winit_state.take_egui_input(&self.window);
-        self.egui_context.begin_frame(input);
-        {
-            let file_epaint_texture_id = self.egui_wgpu_renderer.register_native_texture(
-                &self.device,
-                &self.file_icon_texture.view,
-                wgpu::FilterMode::Linear,
-            );
-
-            let directory_epaint_texture_id = self.egui_wgpu_renderer.register_native_texture(
-                &self.device,
-                &self.directory_icon_texture.view,
-                wgpu::FilterMode::Linear,
-            );
-
-            let play_icon_epaint_texture_id = self.egui_wgpu_renderer.register_native_texture(
-                &self.device,
-                &self.play_icon_texture.view,
-                wgpu::FilterMode::Linear,
-            );
-
-            let mut render_output_epaint_texture_id: Option<egui::epaint::TextureId> = None;
-
-            if self.frame_texture_view.is_some() {
-                render_output_epaint_texture_id =
-                    Some(self.egui_wgpu_renderer.register_native_texture(
-                        &self.device,
-                        &self.frame_texture_view.as_ref().unwrap(),
-                        wgpu::FilterMode::default(),
-                    ));
-            }
-
-            let new_aspect_ratio = dream_editor::render_egui_editor_content(
-                &self.egui_context,
-                render_output_epaint_texture_id,
-                file_epaint_texture_id,
-                directory_epaint_texture_id,
-                play_icon_epaint_texture_id,
-            );
-
-            if self.camera.aspect != new_aspect_ratio {
-                self.camera.aspect = new_aspect_ratio;
-                self.camera.build_view_projection_matrix();
-                self.camera_uniform.update_view_proj(&self.camera);
-                self.queue.write_buffer(
-                    &self.camera_buffer,
-                    0,
-                    bytemuck::cast_slice(&[self.camera_uniform]),
-                );
-            }
-        }
-        let egui_full_output = self.egui_context.end_frame();
-
-        let egui_paint_jobs = self.egui_context.tessellate(egui_full_output.shapes);
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("EGUI Render Encoder"),
-            });
-
-        {
-            for (id, image_delta) in &egui_full_output.textures_delta.set {
-                self.egui_wgpu_renderer
-                    .update_texture(&self.device, &self.queue, *id, image_delta)
-            }
-
-            let egui_screen_descriptor = egui_wgpu::renderer::ScreenDescriptor {
-                size_in_pixels: [self.config.width, self.config.height],
-                pixels_per_point: self.window.scale_factor() as f32,
-            };
-
-            self.egui_wgpu_renderer.update_buffers(
-                &self.device,
-                &self.queue,
-                &mut encoder,
-                &egui_paint_jobs,
-                &egui_screen_descriptor,
-            );
-
-            // draw editor
-            {
-                let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                    label: Some("EGUI Render Pass"),
-                    color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                        view: &view,
-                        resolve_target: None,
-                        ops: wgpu::Operations {
-                            load: wgpu::LoadOp::Clear(wgpu::Color {
-                                r: 0.10588235294,
-                                g: 0.10588235294,
-                                b: 0.10588235294,
-                                a: 1.0,
-                            }),
-                            store: true,
-                        },
-                    })],
-                    depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                        view: &self.depth_texture_egui.view,
-                        depth_ops: Some(wgpu::Operations {
-                            load: wgpu::LoadOp::Clear(1.0),
-                            store: true,
-                        }),
-                        stencil_ops: None,
-                    }),
-                });
-
-                self.egui_wgpu_renderer.render(
-                    &mut render_pass,
-                    &egui_paint_jobs,
-                    &egui_screen_descriptor,
-                );
-            }
-
-            self.queue.submit(iter::once(encoder.finish()));
-            output.present();
-        }
-
-        for id in &egui_full_output.textures_delta.free {
-            self.egui_wgpu_renderer.free_texture(id);
-        }
-
-        Ok(())
-    }
-}
-
-cfg_if::cfg_if! {
-    if #[cfg(target_arch = "wasm32")] {
-        static mut WINDOW_WIDTH: u32 = 2000;
-        static mut WINDOW_HEIGHT: u32 = 1500;
-        static mut NEED_TO_RESIZE_WINDOW: bool = false;
-    }
-}
 
 cfg_if::cfg_if! {
     if #[cfg(target_arch = "wasm32")] {
         #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
         pub fn resize_window(width: u32, height: u32) {
-            unsafe {
-                WINDOW_WIDTH = width;
-                WINDOW_HEIGHT = height;
-                NEED_TO_RESIZE_WINDOW = true;
-            }
+            dream_application::resize_window(width, height);
         }
     }
 }
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen(start))]
 pub async fn run() {
-    cfg_if::cfg_if! {
-        if #[cfg(target_arch = "wasm32")] {
-            std::panic::set_hook(Box::new(console_error_panic_hook::hook));
-            console_log::init_with_level(log::Level::Warn).expect("Could't initialize logger");
-        } else {
-            env_logger::init();
-        }
-    }
-
-    {
-        // example execution of javascript code
-        // let js_code = "console.log('hello world');";
-        let js_code = "7 * 8.1";
-        let mut context = boa_engine::Context::default();
-        match context.eval(js_code) {
-            Ok(res) => {
-                println!("{}", res.to_string(&mut context).unwrap());
-                log::warn!("{}", res.to_string(&mut context).unwrap());
-            }
-            Err(e) => {
-                // Pretty print the error
-                eprintln!("Uncaught {}", e.display());
-                log::error!("Uncaught {}", e.display());
-            }
-        };
-    }
-
-    let event_loop = EventLoop::new();
-    let window = WindowBuilder::new()
-        .with_inner_size(winit::dpi::PhysicalSize::new(3000, 1750))
-        .build(&event_loop)
-        .unwrap();
-
-    cfg_if::cfg_if! {
-        if #[cfg(target_arch = "wasm32")] {
-        } else {
-            window.set_title("Dream Engine");
-        }
-    }
-
-    #[cfg(target_arch = "wasm32")]
-    {
-        // Winit prevents sizing with CSS, so we have to set
-        // the size manually when on web.
-        window.set_inner_size(winit::dpi::PhysicalSize::new(2000, 1500));
-
-        use winit::platform::web::WindowExtWebSys;
-        web_sys::window()
-            .and_then(|win| win.document())
-            .and_then(|doc| {
-                let dst = doc.get_element_by_id("wasm-example")?;
-                let canvas = web_sys::Element::from(window.canvas());
-                dst.append_child(&canvas).ok()?;
-                Some(())
-            })
-            .expect("Couldn't append canvas to document body.");
-    }
-
-    // State::new uses async code, so we're going to wait for it to finish
-    let mut state = State::new(window, &event_loop).await;
-    event_loop.run(move |event, _, control_flow| {
-        #[cfg(target_arch = "wasm32")]
-        {
-            unsafe {
-                if NEED_TO_RESIZE_WINDOW {
-                    state
-                        .window()
-                        .set_inner_size(winit::dpi::PhysicalSize::new(WINDOW_WIDTH, WINDOW_HEIGHT));
-                    state.resize(winit::dpi::PhysicalSize::new(WINDOW_WIDTH, WINDOW_HEIGHT));
-                    NEED_TO_RESIZE_WINDOW = false;
-                }
-            }
-        }
-        match event {
-            Event::RedrawRequested(window_id) if window_id == state.window().id() => {
-                state.update();
-
-                match state.render() {
-                    Ok(_) => {}
-                    // Reconfigure the surface if it's lost or outdated
-                    Err(wgpu::SurfaceError::Lost | wgpu::SurfaceError::Outdated) => {
-                        state.resize(state.size)
-                    }
-                    // The system is out of memory, we should probably quit
-                    Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                    // We're ignoring timeouts
-                    Err(wgpu::SurfaceError::Timeout) => log::warn!("Surface timeout"),
-                }
-
-                match state.render_egui() {
-                    Ok(_) => {}
-                    // Reconfigure the surface if it's lost or outdated
-                    Err(wgpu::SurfaceError::Lost | wgpu::SurfaceError::Outdated) => {
-                        state.resize(state.size)
-                    }
-                    // The system is out of memory, we should probably quit
-                    Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                    // We're ignoring timeouts
-                    Err(wgpu::SurfaceError::Timeout) => log::warn!("Surface timeout"),
-                }
-            }
-            Event::WindowEvent { event, .. } => {
-                let exclusive = state.egui_winit_state.on_event(&state.egui_context, &event);
-                if !exclusive.consumed {
-                    match event {
-                        WindowEvent::Resized(physical_size) => {
-                            state.resize(physical_size);
-                        }
-                        WindowEvent::CloseRequested => *control_flow = ControlFlow::Exit,
-                        WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
-                            // new_inner_size is &mut so w have to dereference it twice
-                            state.resize(*new_inner_size);
-                        }
-                        _ => (),
-                    }
-                }
-            }
-            Event::MainEventsCleared => {
-                // RedrawRequested will only trigger once, unless we manually
-                // request it.
-                state.window().request_redraw();
-            }
-            _ => {}
-        }
-    });
+    dream_application::run().await;
 }
